move converter to utils
