+++ logzio-lambda-extensions-logs/utils/converter.go
@@ -1,4 +1,4 @@
-package logs_converter
+package utils
 
 const (
 	fldLogzioTimestamp = "@timestamp"
